Gateway added - now this is follow service only

+++ src/main/java/com/tripshorts/userservice/services/implementation/FollowServiceImplementation.java
@@ -2,17 +2,15 @@ package com.tripshorts.userservice.services.implementation;
 
 import com.tripshorts.userservice.entity.FollowersTable;
 import com.tripshorts.userservice.entity.UserEntity;
-import com.tripshorts.userservice.exceptions.UserNotFound;
+import com.tripshorts.userservice.exceptions.UserException;
 import com.tripshorts.userservice.model.UserDTO;
 import com.tripshorts.userservice.repositeries.FollowRepository;
 import com.tripshorts.userservice.repositeries.UserRepository;
 import com.tripshorts.userservice.services.FollowService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
-import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -27,15 +25,18 @@ public class FollowServiceImplementation implements FollowService {
     private UserRepository userRepository;
 
     @Override
-    public UserDTO follow(String username, String currUsername) throws UserNotFound {
+    public UserDTO follow(String username, String currUserId) throws UserException {
         Optional<UserEntity> userEntity = userRepository.findByUsername(username);
-        Optional<UserEntity> currentUserEntity = userRepository.findByUsername(currUsername);
+        Optional<UserEntity> currentUserEntity = userRepository.findById(Long.parseLong(currUserId));
         if(!userEntity.isPresent() || !currentUserEntity.isPresent())
-            throw new UserNotFound("User Not Found");
+            throw new UserException("User Not Found");
         try {
+            FollowersTable resFollowersTable = followRepository.findByUserEntityAndUserFollowingEntity(userEntity.get(), currentUserEntity.get());
+            if(resFollowersTable != null)
+                throw new UserException("Already following user with username: "+username);
             FollowersTable followersTable = new FollowersTable();
             followersTable.setUserEntity(currentUserEntity.get());
-            followersTable.setUserFollowerEntity(userEntity.get());
+            followersTable.setUserFollowingEntity(userEntity.get());
             followRepository.save(followersTable);
             UserDTO userDTO = new UserDTO();
             BeanUtils.copyProperties(userEntity, userDTO);
@@ -46,14 +47,16 @@ public class FollowServiceImplementation implements FollowService {
     }
 
     @Override
-    public UserDTO unfollow(String username, String currUsername) throws UserNotFound {
+    public UserDTO unfollow(String username, String userId) throws UserException {
         Optional<UserEntity> userEntity = userRepository.findByUsername(username);
-        Optional<UserEntity> currentUserEntity = userRepository.findByUsername(currUsername);
+        Optional<UserEntity> currentUserEntity = userRepository.findById(Long.parseLong(userId));
 
         if(!userEntity.isPresent() || !currentUserEntity.isPresent())
-            throw new UserNotFound("User Not Found");
+            throw new UserException("User Not Found");
         try {
-            FollowersTable followersTable = followRepository.findByUserEntityAndUserFollowerEntity(currentUserEntity.get(), userEntity.get());
+            FollowersTable followersTable = followRepository.findByUserEntityAndUserFollowingEntity(currentUserEntity.get(), userEntity.get());
+            if(followersTable == null)
+                throw new UserException("Not following user with username: "+ username);
             followRepository.delete(followersTable);
             UserDTO userDTO = new UserDTO();
             BeanUtils.copyProperties(userEntity, userDTO);
@@ -64,11 +67,11 @@ public class FollowServiceImplementation implements FollowService {
     }
 
     @Override
-    public List<UserDTO> getUserFollowers(String username) throws UserNotFound {
-        Optional<UserEntity> userEntity = userRepository.findByUsername(username);
+    public List<UserDTO> getUserFollowers(String userId) throws UserException {
+        Optional<UserEntity> userEntity = userRepository.findById(Long.parseLong(userId));
         if(!userEntity.isPresent())
-            throw new UserNotFound("Unable to retrieve followers: User not found");
-        List<FollowersTable> followersTableList = followRepository.findByUserFollowerEntity(userEntity.get());
+            throw new UserException("Unable to retrieve followers: User not found");
+        List<FollowersTable> followersTableList = followRepository.findByUserFollowingEntity(userEntity.get());
         List<UserDTO> userDTOs = new ArrayList<UserDTO>();
         for(FollowersTable followersTableEntry: followersTableList) {
             UserDTO userDTO = new UserDTO();
@@ -79,22 +82,18 @@ public class FollowServiceImplementation implements FollowService {
     }
 
     @Override
-    public List<UserDTO> getUserFollowing(String username) throws UserNotFound {
-        Optional<UserEntity> userEntity = userRepository.findByUsername(username);
+    public List<UserDTO> getUserFollowing(String userId) throws UserException {
+        Optional<UserEntity> userEntity = userRepository.findById(Long.parseLong(userId));
         if(!userEntity.isPresent())
-            throw new UserNotFound("Unable to retrieve followers: User not found");
+            throw new UserException("Unable to retrieve following: User not found");
         List<FollowersTable> followersTableList = followRepository.findByUserEntity(userEntity.get());
         List<UserDTO> userDTOs = new ArrayList<UserDTO>();
         for(FollowersTable followersTableEntry: followersTableList) {
             UserDTO userDTO = new UserDTO();
-            BeanUtils.copyProperties(followersTableEntry.getUserFollowerEntity(), userDTO);
+            BeanUtils.copyProperties(followersTableEntry.getUserFollowingEntity(), userDTO);
             userDTOs.add(userDTO);
         }
         return userDTOs;
     }
 
-//    @Override
-//    public List<Long> getUserById(Long id) {
-//        return followRepository.findById(id);
-//    }
 }
